plugin now correctly removes any constant that appears in nvim's c.vim

+++ scripts/extrasyntax.sh
@@ -8,16 +8,16 @@ CPP=$(which cpp)
 DIFF=$(which diff)
 
 function find_constants () {
-    [[ -z "${1}" || ! -f "${1}" ]] && echo "" && return 1
-    [[ ! -f "${2}/predefs.txt" ]] && echo "" && return 1
-
     FILE=${1}
     CACHE_DIR=${2}
+
+    [[ -z "${FILE}" || ! -f "${FILE}" || ! -f "${CACHE_DIR}/predefs.txt" ]] && return 1
+
     TMPFILE=$(mktemp)
 
     ${CPP} -dD -E -x c ${FILE} 2>/dev/null | ${GREP} "#define" | cut -d' ' -f2 | grep -v -E -e "\w+\(.*\)$" > ${TMPFILE}
 
-    grep ${TMPFILE} -v -f "${CACHE_DIR}/predefs.txt" | grep -v -E -e "(^_.*|_H(_INCLUDED)?$)"
+    grep ${TMPFILE} -v -f "${CACHE_DIR}/predefs.txt" | grep -v -f "${CACHE_DIR}/nvim_syntax.txt" | grep -v -E -e "(^_.*|_H(_INCLUDED)?$)"
 
     rm ${TMPFILE}
